tests: require kernel-modules to workaround RHBZ#1652823
There is no kernel-debug-modules in rawhide, so the tests are passing there
nicely. However, when the update is submitted to a stable release, the problem
described in RHBZ#1652823 triggers installation of kernel-debug-modules, thus
the test fails due to missing squashfs support.

The recent update of snapd 2.45 provides a nice example of the failure scenario:
https://bodhi.fedoraproject.org/updates/FEDORA-2020-3b4972cd67

Signed-off-by: Maciek Borzecki <maciek.borzecki@gmail.com>